refact: corrige senhas expostas

+++ easymqtt/clearsession.php
@@ -13,8 +13,17 @@ if (!isset($_GET['session']) || empty($_GET['session'])) {
 
 $session = htmlspecialchars($_GET["session"]);
 
-// Conexão direta ao MongoDB via driver nativo
-$manager = new MongoDB\Driver\Manager("mongodb+srv://ti:HjrjfpzWT4cdDJqc@bipes-db.wlo1lu9.mongodb.net/?retryWrites=true&w=majority&appName=bipes-db");
+// Pega a URI do Mongo a partir da variável de ambiente
+$mongoUri = getenv('MONGO_URI');
+if (!$mongoUri) {
+    echo json_encode([
+        "success" => false,
+        "result"  => "Mongo URI is not set in environment"
+    ]);
+    exit;
+}
+
+$manager = new MongoDB\Driver\Manager($mongoUri);
 
 // Comando para remover (dropar) todo o banco de dados da sessão
 $cmd = new MongoDB\Driver\Command(['dropDatabase' => 1]);
